+++ example/lib/src/main/grid_app_example/inherited/view/inherit_cat.dart
@@ -21,8 +21,8 @@ class InheritCat extends StatefulWidget {
 }
 
 /// Supply an InheritedWidget to a StateX object: InheritedStateX
-class _InheritCatState extends StateIn<InheritCat> {
-  _InheritCatState() : super(controller: CatController());
+class _InheritCatState extends StateX<InheritCat> {
+  _InheritCatState() : super(controller: CatController(), useInherited: true);
 
   /// Use buildIn() or builder() instead of build() to call the built-in InheritedWidget
   @override