+++ example/CHANGELOG.md
@@ -1,4 +1,8 @@
 
+## 1.2.0
+- *Deprecated* StateIn() and StateF() class are deprecated to keep it simple.
+  Use the StateX parameters, *runAsync* and *useInherited*, instead.'
+
 ## 1.1.1
 November 29, 2023
 - Relative paths now in import statements.